+++ test/fixtures/golden/h2o2/cuda/rxn_rates_pres_mod.cu
@@ -18,7 +18,7 @@ __device__ void get_rxn_pres_mod (const double T, const double pres, const doubl
   register double B;
 
   register double logT = log(T);
-  register double m = pres / (8.31446210e+03 * T);
+  register double m = pres / (8.31446262e+03 * T);
 
   // reaction 0;
   shared_temp[threadIdx.x + 3 * blockDim.x] = C[INDEX(0)];
@@ -41,7 +41,7 @@ __device__ void get_rxn_pres_mod (const double T, const double pres, const doubl
 
   // reaction 20;
   thd = m + 1.0 * shared_temp[threadIdx.x + 3 * blockDim.x] + 5.0 * shared_temp[threadIdx.x + 2 * blockDim.x] - 0.30000000000000004 * shared_temp[threadIdx.x + 1 * blockDim.x];
-  k0 = exp(2.8463930238863654e+01 - 0.9 * logT - (-8.5547326026057669e+02 / T));
+  k0 = exp(2.8463930238863654e+01 - 0.9 * logT - (-8.5547320694790187e+02 / T));
   kinf = exp(2.5027330930150580e+01 - 0.37 * logT);
   Pr = k0 * thd / kinf;
   logFcent = log10( fmax(2.65400000e-01 * exp(-T / 9.40000000e+01) + 7.34600000e-01 * exp(-T / 1.75600000e+03) + exp(-5.18200000e+03 / T), 1.0e-300));